Refactor embedding of Unicode data - decrease execution time
Why not to use golang.org/x/text/unicode: The built binary was larger
than 3.1mbyte in size. In addition, benchmarking the execution with
hyperfine revealed 200ms± execution time

    hyperfine --warmup 5 './bin/ucn-1.0.0 -filter.starts-with math -filter.starts-with pi'
    Time (mean ± σ):     212.6 ms ±   4.2 ms    [User: 210.7 ms, System: 8.9 ms]
    Range (min … max):   208.0 ms … 222.9 ms    13 runs

One of the reasons is the fact that golang.org/x/text/unicode/runenames
contains 130k names in the form `<Private Use>` and another 100k entries
like `<CJK Ideograph Extension...`. So, out of the 288767 names in
`ucn:1.0.0`, only 40k are directly useful. This commit filters out not
directly useful names via `make generate-internals`. With this, we
can get rid of golang.org/x/text/unicode, improve usability and
code execution.

Why not to use a regular data structure like a slice? The resulting
binary is 5mb, which is the biggest. On the other hand, it is pretty
fast:

    hyperfine --warmup 5 './bin/ucn.slice -filter.starts-with math -filter.starts-with pi'
    Time (mean ± σ):      38.8 ms ±   6.0 ms    [User: 36.8 ms, System: 3.2 ms]
    Range (min … max):    36.3 ms …  77.2 ms    73 runs

Why not zstd + b85? The data itself _is_ actually smaller
(gzip-9: 244kb, zstd-19: 144kb). But, to decompress the data requires a
zstd-decompressor (github.com/klauspost/compress/zstd is awesome), which
itself adds more than the 100kb delta.

The picked gzip+b64 approach yields a binary size of 2.8mb and an execution time
of 50ms± (a 4x decrease in comparison to the golang.org/x/text/unicode
approach):

    hyperfine --warmup 5 './bin/ucn -filter.starts-with math -filter.starts-with pi'
    Time (mean ± σ):      45.1 ms ±   0.9 ms    [User: 46.7 ms, System: 3.5 ms]
    Range (min … max):    43.4 ms …  47.5 ms    61 runs