wip: add print info about code points
The initial use case of `ucn` was to search for a given term in the
description of the available Unicode characters to pick the desired
character or the hex-code etc. and use it in either prose or code.

Every now and then, the reverse use case occurs: one spots a unicode
digit in code and want to know the name of the referenced character.

This commit adds this feature: any reguler, non-flag argument given
to `ucn` is treated as filter for the whole body of embedded Unicode
characters. Example given:

    ucn 00a1

yields

    '¡'    00a1     U+00A1     &iexcl;     \\u00a1     INVERTED EXCLAMATION MARK
    '𐂡'    100a1    U+100A1    &#100a1;    \\u100a1    LINEAR B IDEOGRAM B152

The same output is produced via `ucn '¡'`.